python/optionslib/portfolio.py
"""Portfolio management module."""

import numpy as np
from typing import Dict, List, Optional, Tuple, Union

class Portfolio:
    """Python implementation of the Portfolio class."""
    
    def __init__(self):
        """Initialize the portfolio."""
        self._positions = {}
        
    def add_position(self, instrument_id: str, quantity: float, 
                    pricing_model: str, params: Dict[str, float]) -> None:
        """Add a position to the portfolio."""
        self._positions[instrument_id] = {
            'quantity': quantity,
            'pricing_model': pricing_model,
            'params': params
        }
        
    def remove_position(self, instrument_id: str) -> None:
        """Remove a position from the portfolio."""
        if instrument_id in self._positions:
            del self._positions[instrument_id]
            
    def get_position(self, instrument_id: str) -> Optional[Dict]:
        """Get position details."""
        return self._positions.get(instrument_id)
    
    def get_all_positions(self) -> Dict:
        """Get all positions."""
        return self._positions.copy()
    
    def calculate_portfolio_value(self, market_data: Dict) -> float:
        """Calculate the total portfolio value."""
        total_value = 0.0
        for instrument_id, position in self._positions.items():
            if instrument_id in market_data:
                price = market_data[instrument_id]
                total_value += position['quantity'] * price
        return total_value
    
    def calculate_portfolio_risk(self, market_data: Dict, risk_metric: str = 'var') -> float:
        """Calculate portfolio risk using specified risk metric."""
        # Implement basic risk calculations
        if risk_metric.lower() == 'var':
            # Simple Value at Risk calculation
            portfolio_values = []
            for _ in range(1000):  # Monte Carlo simulation
                simulated_value = self.calculate_portfolio_value(market_data)
                portfolio_values.append(simulated_value)
            return np.percentile(portfolio_values, 5)  # 95% VaR
        return 0.0
